Fix two instances of Clang's -Wdouble-promotion
src/frontends/qt/GuiView.cpp:184:30: warning: implicit conversion increases floating-point precision: 'float' to 'qreal' (aka 'double') [-Wdouble-promotion]
  184 |                 qreal locscale = htextsize.toFloat(&ok);
      |                       ~~~~~~~~   ~~~~~~~~~~^~~~~~~~~~~~

src/frontends/qt/FancyLineEdit.cpp:344:22: warning: implicit conversion increases floating-point precision: 'float' to 'qreal' (aka 'double') [-Wdouble-promotion]
  344 |                 painter.setOpacity(m_iconOpacity);
      |                         ~~~~~~~~~~ ^~~~~~~~~~~~~

+++ src/frontends/qt/FancyLineEdit.h
@@ -25,7 +25,7 @@ class FancyLineEditPrivate;
 class IconButton: public QAbstractButton
 {
 	Q_OBJECT
-	Q_PROPERTY(float iconOpacity READ iconOpacity WRITE setIconOpacity)
+	Q_PROPERTY(qreal iconOpacity READ iconOpacity WRITE setIconOpacity)
 	Q_PROPERTY(bool autoHide READ hasAutoHide WRITE setAutoHide)
 	Q_PROPERTY(QPixmap pixmap READ pixmap WRITE setPixmap)
 public:
@@ -33,14 +33,14 @@ public:
 	void paintEvent(QPaintEvent *event) override;
 	void setPixmap(const QPixmap &pixmap) { m_pixmap = pixmap; update(); }
 	QPixmap pixmap() const { return m_pixmap; }
-	float iconOpacity() { return m_iconOpacity; }
-	void setIconOpacity(float value) { m_iconOpacity = value; update(); }
+	qreal iconOpacity() { return m_iconOpacity; }
+	void setIconOpacity(qreal value) { m_iconOpacity = value; update(); }
 	void animateShow(bool visible);
 
 	void setAutoHide(bool hide) { m_autoHide = hide; }
 	bool hasAutoHide() const { return m_autoHide; }
 private:
-	float m_iconOpacity;
+	qreal m_iconOpacity;
 	bool m_autoHide;
 	QPixmap m_pixmap;
 };
